refactor(examples): merge structure into pages
structure.ts had no independent consumer and no tests protecting its purity, so
fold the tree model (types, build, leaves) into pages.ts. The whole pages
concern — globs, layout conventions, tree, iframe protocol — now lives in one
file; leaf-file wiring is inlined into the build instead of a follow-up loop.
